+++ temp-seed-scripts/vega-migration/seed-perception-data/index.js
@@ -5,6 +5,8 @@ require('./../../../src/backend/auth0/vegaClient')
 
 const zero_resetPerceptionData = require('./zero_resetPerceptionData')
 const one_IndRegProdSteps = require('./one_IndRegProdSteps')
+const two_clientTeamSteps = require('./two_clientTeamSteps')
+
 /*
   Migration steps:
     0. Delete all relevant data from vega
@@ -12,6 +14,9 @@ const one_IndRegProdSteps = require('./one_IndRegProdSteps')
       a. Seed Regimens
       b. Seed Products and seed ProdReg join table through pulse-core.regimens collection
       c. Seed Indicsations and Seed IndReg connections through pulse-core.regimens collection
+    2. Client-Team Steps
+      a. Seed Clients
+      b. Seed Teams w/ client foreign keys
 */
 
 const migrateToVega = async () => {
@@ -21,6 +26,7 @@ const migrateToVega = async () => {
 
   await zero_resetPerceptionData()
   await one_IndRegProdSteps(dbs)
+  await two_clientTeamSteps(dbs)
 
   console.log('Perception Tool Vega Migration Completed')
   dbs.close()